requirements.txt
django 
djangorestframework
piper-tts>=0.14.5
pymysql>=1.1